+++ CHANGES.md
@@ -1,5 +1,6 @@
 Version 1.0.7-dev
 -----------------
+- Add `best_epoch` attribute to `SaveBestWeights` and `KeepBestWeights`.
 
 
 Version 1.0.6 [12 Jun 2026]